# Documentation/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Birøkt API
  description: |
    RESTful API for birøkt-systemet (beekeeping management system).
    
    ## Authentication
    All endpoints (except auth endpoints) require a Bearer token in the Authorization header.
    
    ## Rate Limiting
    - Auth endpoints: 5 requests/minute per IP
    - Read endpoints: 100 requests/minute per user
    - Write endpoints: 30 requests/minute per user
    
  version: 1.0.0
  contact:
    name: Birøkt Support
    email: support@birokt.no
  license:
    name: Proprietary
servers:
  - url: https://api.birokt.no/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Inspection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hive:
          type: object
          properties:
            id:
              type: string
              format: uuid
            hive_number:
              type: string
            apiary_name:
              type: string
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        inspection_date:
          type: string
          format: date-time
        weather:
          type: object
          properties:
            temperature:
              type: number
              format: float
            wind_speed:
              type: number
              format: float
            condition:
              type: string
        assessment:
          type: object
          properties:
            strength:
              type: string
              enum: [weak, medium, strong]
            temperament:
              type: string
              enum: [calm, nervous, aggressive]
            queen_seen:
              type: boolean
            queen_laying:
              type: boolean
        frames:
          type: object
          properties:
            brood:
              type: integer
            honey:
              type: integer
            pollen:
              type: integer
            empty:
              type: integer
        health:
          type: object
          properties:
            status:
              type: string
              enum: [healthy, warning, critical]
            varroa_level:
              type: string
              enum: [none, low, medium, high]
            diseases:
              type: array
              items:
                type: string
            pests:
              type: array
              items:
                type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Photo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        file_size:
          type: integer
        mime_type:
          type: string
        width:
          type: integer
        height:
          type: integer
        caption:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /inspections:
    get:
      tags:
        - Inspections
      summary: Get inspections
      security:
        - bearerAuth: []
      parameters:
        - name: hive_id
          in: query
          schema:
            type: string
            format: uuid
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of inspections
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inspection'

    post:
      tags:
        - Inspections
      summary: Create new inspection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hive_id
                - inspection_date
              properties:
                hive_id:
                  type: string
                  format: uuid
                inspection_date:
                  type: string
                  format: date-time
                weather:
                  type: object
                assessment:
                  type: object
                frames:
                  type: object
                health:
                  type: object
                notes:
                  type: string
      responses:
        '201':
          description: Inspection created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Inspection'
...